sets connection gater and peer manager default
