+++ cmd/util/cmd/execution-state-extract/export_report.json
@@ -1,6 +1,6 @@
 {
  "EpochCounter": 0,
- "PreviousStateCommitment": "e5c4341c12bd78c2dcc0ae874fed4525d27c3328aa30cd338dff99ceabe786d0",
- "CurrentStateCommitment": "e5c4341c12bd78c2dcc0ae874fed4525d27c3328aa30cd338dff99ceabe786d0",
+ "PreviousStateCommitment": "88f5861f6d3003c5b449343fcac1fe433d167f36392f13d6a4ceb62bd093b741",
+ "CurrentStateCommitment": "88f5861f6d3003c5b449343fcac1fe433d167f36392f13d6a4ceb62bd093b741",
  "ReportSucceeded": true
 }
