+++ network/p2p/p2pnode/disallow_listing_test.go
@@ -11,7 +11,6 @@ import (
 	"github.com/onflow/flow-go/module/irrecoverable"
 	mockmodule "github.com/onflow/flow-go/module/mock"
 	"github.com/onflow/flow-go/network"
-	"github.com/onflow/flow-go/network/internal/testutils"
 	"github.com/onflow/flow-go/network/p2p"
 	"github.com/onflow/flow-go/network/p2p/connection"
 	p2pconfig "github.com/onflow/flow-go/network/p2p/p2pbuilder/config"
@@ -44,7 +43,7 @@ func TestDisconnectingFromDisallowListedNode(t *testing.T) {
 			func() peer.IDSlice {
 				return peerIDSlice
 			}),
-		p2ptest.WithConnectionGater(testutils.NewConnectionGater(idProvider, func(p peer.ID) error {
+		p2ptest.WithConnectionGater(p2ptest.NewConnectionGater(idProvider, func(p peer.ID) error {
 			// allow all the connections, except for the ones that are disallow-listed, which are determined when
 			// this connection gater object queries the disallow listing oracle that will be provided to it by
 			// the libp2p node. So, here, we don't need to do anything except just enabling the connection gater.
